% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelioration-annotation.R
\name{amelioration.annotation}
\alias{amelioration.annotation}
\title{Fonction qui retire les underscore dans la colonne gene des annotation}
\usage{
amelioration.annotation(species, collicin, annotationDir)
}
\arguments{
\item{species}{espece consernée}

\item{collicin}{les gènes consernés}

\item{annotationDir}{la localisation des fichiers d'annotations}
}
\value{

}
\description{
Fonction qui retire les underscore dans la colonne gene des annotation
}
