% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-annotation.R
\name{analyse.annotation}
\alias{analyse.annotation}
\title{A function to summarize the presence of some genes based on annotation}
\usage{
analyse.annotation(bacteria.table, collicin, annotationDir)
}
\arguments{
\item{bacteria.table}{: the bacteria table}

\item{collicin}{the collicin data.frame}

\item{annotationDir}{the directory where the annotation list can be found}
}
\value{

}
\description{
A function to summarize the presence of some genes based on annotation
}
