% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-blasttable.R
\name{analyse.blast.table}
\alias{analyse.blast.table}
\title{A function to summarize the presence of some genes based on blast table}
\usage{
analyse.blast.table(bacteria.table, blastresult)
}
\arguments{
\item{bacteria.table}{: the bacteria table}

\item{blastresult}{the blast result data.frame}
}
\value{

}
\description{
A function to summarize the presence of some genes based on blast table
}
