% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.n.bacteria.R
\name{compute.n.bacteria}
\alias{compute.n.bacteria}
\title{Compute the number of genome available for each bacteria}
\usage{
compute.n.bacteria(bacteria.table)
}
\arguments{
\item{bacteria.table}{a data.frame produce by the function extract.bacteria.table}
}
\value{
a table with the number of available genomes for each bacteria
}
\description{
to compute later
}
