% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.annot.genome.R
\name{dl.annot.genome}
\alias{dl.annot.genome}
\title{telecharge les fichiers d'annotations et des genomes}
\usage{
dl.annot.genome(
  species,
  NmaxPlasmid = 1000,
  annotationDir,
  genomeDir,
  accessionDir
)
}
\arguments{
\item{accessionDir}{}
}
\value{

}
\description{
telecharge les fichiers d'annotations et des genomes
}
