% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.accession.package.R
\name{download.accession.chromosome.plasmid}
\alias{download.accession.chromosome.plasmid}
\title{Create accession csv file à la base du package genome. il sort les accessions numbers des chromosomes et des plasmides}
\usage{
download.accession.chromosome.plasmid(species, bacteria.table, outDir)
}
\arguments{
\item{species}{a vector of character including the species that we want to analyze}

\item{bacteria.table}{The bacteria table produced by the create.bacteria.table function}

\item{outDir}{The output directory}
}
\value{
Ecrit dans le repertoire un fichier csv contenant les accessions listes
}
\description{
Create accession csv file à la base du package genome. il sort les accessions numbers des chromosomes et des plasmides
}
