% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.accession.NCBI.chromosome.R
\name{download.accession.NCBI.chromosome}
\alias{download.accession.NCBI.chromosome}
\title{Telecharge les accessions list sur NCBI}
\usage{
download.accession.NCBI.chromosome(accessionDir)
}
\arguments{
\item{accessionDir}{}
}
\value{

}
\description{
Telecharge les accessions list sur NCBI
}
