% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.annotation.R
\name{download.annotation}
\alias{download.annotation}
\title{Telecharge les annotations}
\usage{
download.annotation(species, indextostart, accessionDir, outDir)
}
\arguments{
\item{species}{a vector of character including the species that we want to analyze}

\item{indextostart}{: the index to restart the analysis. for example, if you interupted the analysis after 20 species, you can specify indextorestart=21}

\item{accessionDir}{the directory where the accession list can be found}

\item{outDir}{The output directory}
}
\value{
un fichier par souche contenant les annotations dans le repertoire de sortie
}
\description{
Telecharge les annotations
}
