% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-annotation.R
\name{downloadannotation}
\alias{downloadannotation}
\title{Title}
\usage{
downloadannotation(maxOrganism = 20, accession.list, repertoire)
}
\arguments{
\item{repertoire}{}
}
\value{

}
\description{
Title
}
