% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-genome.R
\name{downloadgenome}
\alias{downloadgenome}
\title{Telecharge le genome}
\usage{
downloadgenome(maxOrganism, accession.list, repertoire)
}
\arguments{
\item{maxOrganism}{un entier du nombre d'organisme max}

\item{accession.list}{la liste des différentes especes}

\item{repertoire}{l'emplacement pour telecharger les fichiers}
}
\value{

}
\description{
Telecharge le genome
}
