% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.bacteria.table.R
\name{extract.bacteria.table}
\alias{extract.bacteria.table}
\title{extract a table of bacteria with accession numbers based on the genome package}
\usage{
extract.bacteria.table(phylum)
}
\arguments{
\item{phylum}{a data.frame that can be produced by the function extract.phylum}
}
\value{
a table with the one line for each genome available
}
\description{
to compute later
}
