% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.phylum.table.R
\name{extract.phylum}
\alias{extract.phylum}
\title{extract a table of the phylum of bacteria based on the taxize package}
\usage{
extract.phylum()
}
\value{
a table with the one line for each phylum
}
\description{
to compute later
}
