% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.sequence.nom.R
\name{extract.sequences.annotation.nom}
\alias{extract.sequences.annotation.nom}
\title{extrait les sequences comportant colicin et ne comportant pas resistance,immunity et transporter dans le produit des annotations utilisé pour les key}
\usage{
extract.sequences.annotation.nom(species, annotationDir, outDir, genomeDir)
}
\arguments{
\item{genomeDir}{}
}
\value{

}
\description{
extrait les sequences comportant colicin et ne comportant pas resistance,immunity et transporter dans le produit des annotations utilisé pour les key
}
