% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.sequences.annotation.key.R
\name{extract.sequences.annotation}
\alias{extract.sequences.annotation}
\title{extrait les sequences comportant colicin et ne comportant pas resistance et immunity dans le produit des annotations utilisé pour les key}
\usage{
extract.sequences.annotation(species, annotationDir, outDir, genomeDir)
}
\arguments{
\item{genomeDir}{}
}
\value{

}
\description{
extrait les sequences comportant colicin et ne comportant pas resistance et immunity dans le produit des annotations utilisé pour les key
}
