% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gram.R
\name{Gram}
\alias{Gram}
\title{ajoute si le sous-groupe fait partie des bactéries gram + ou -}
\usage{
Gram(presence.summary.subgroup)
}
\arguments{
\item{presence.summary.subgroup}{}
}
\value{

}
\description{
ajoute si le sous-groupe fait partie des bactéries gram + ou -
}
