% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.all.R
\name{phylo.all}
\alias{phylo.all}
\title{fait un arbre phylogenetic. Utiliser avec de grand nombre de sequences}
\usage{
phylo.all(bacteria.table, Dir)
}
\arguments{
\item{Dir}{}
}
\value{

}
\description{
fait un arbre phylogenetic. Utiliser avec de grand nombre de sequences
}
