% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo.from.otus.R
\name{phylo.from.otus}
\alias{phylo.from.otus}
\title{A function to plot phylogenetic tree of all genes based on fasta file in the consensus directory}
\usage{
phylo.from.otus(otusDir, bacteria.table)
}
\arguments{
\item{otusDir}{the directory with the consensus}

\item{bacteria.table}{bacteria table}
}
\value{

}
\description{
A function to plot phylogenetic tree of all genes based on fasta file in the consensus directory
}
