% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylo.R
\name{phylo}
\alias{phylo}
\title{fait un arbre phylogenetic. Utiliser avec de grand nombre de sequences}
\usage{
phylo(sequences)
}
\arguments{
\item{bacteria.table}{}
}
\value{

}
\description{
fait un arbre phylogenetic. Utiliser avec de grand nombre de sequences
}
