% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.blast.R
\name{screenBlast}
\alias{screenBlast}
\title{Screen a sequence using Blast}
\usage{
screenBlast(reference, querry, min.pc.ident, min.pc.length)
}
\arguments{
\item{reference}{the reference sequence that you want to screen. Fasta file in one or severa sequences}

\item{querry}{the querry sequence. Fasta file in one or severa sequences}

\item{min.pc.length}{}

\item{dir.out}{the directory of the output}
}
\value{
numeric value of the percentage of the reference sequence which is covered by the querry
}
\description{
You provide a reference and a querry and the function compute the percentage of the reference which is covered by the querry
}
