% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.blastkey.R
\name{screenBlastkey}
\alias{screenBlastkey}
\title{Screen a sequence using Blast}
\usage{
screenBlastkey(reference, querry, min.pc.ident, min.pc.length, geneDir)
}
\arguments{
\item{reference}{the reference sequence that you want to screen. Fasta file in one or severa sequences}

\item{querry}{the querry sequence. Fasta file in one or severa sequences}

\item{geneDir}{}
}
\value{
numeric value of the percentage of the reference sequence which is covered by the querry and write dans le geneDir
la sequences des gènes similaires
}
\description{
You provide a reference and a querry and the function compute the percentage of the reference which is covered by the querry
}
