% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writteextractallgenes.R
\name{writteextractallgenes}
\alias{writteextractallgenes}
\title{Title}
\usage{
writteextractallgenes(
  accession.list,
  collicin,
  repertoire__gene_annotation,
  repertoire_annotation,
  repertoire_genome
)
}
\arguments{
\item{repertoire__gene_annotation}{}
}
\value{

}
\description{
Title
}
