% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.accession.bacteriocin.R
\name{download.accession.bacteriocin}
\alias{download.accession.bacteriocin}
\title{Telecharge les accessions de bacteriocin sur NCBI}
\usage{
download.accession.bacteriocin(outDir)
}
\arguments{
\item{outDir}{The output directory}
}
\value{
Ecrit dans le repertoire un fichier csv contenant les accessions listes
}
\description{
Telecharge les accessions de bacteriocin sur NCBI
}
