% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.accession.chromosome.R
\name{download.accession.chromosome}
\alias{download.accession.chromosome}
\title{Telecharge les accessions list des chromosomes 1 et 2 de vibrio cholerae sur NCBI}
\usage{
download.accession.chromosome(outDir)
}
\arguments{
\item{outDir}{output directory}
}
\value{
Ecrit dans le repertoire un fichier csv contenant les accessions listes
}
\description{
Telecharge les accessions list des chromosomes 1 et 2 de vibrio cholerae sur NCBI
}
