% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.accession.chromosome.R
\name{download.accession.chromosome}
\alias{download.accession.chromosome}
\title{Title Create accession csv file}
\usage{
download.accession.chromosome(species, bacteria.table, outDir)
}
\arguments{
\item{species}{a vector of character including the species that we want to analyze}

\item{bacteria.table}{The bacteria table produced by the create.bacteria.table function}

\item{outDir}{The output directory}
}
\value{

}
\description{
Title Create accession csv file
}
