% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.annotation.R
\name{download.annotation}
\alias{download.annotation}
\title{Title}
\usage{
download.annotation(
  species,
  maxOrganism = 20,
  indextostart,
  accessionDir,
  outDir
)
}
\arguments{
\item{species}{a vector of character including the species that we want to analyze}

\item{maxOrganism}{a value with the number of genome we want to download for each species}

\item{indextostart}{: the index to restart the analysis. for example, if you interupted the analysis after 20 species, you can specify indextorestart=21}

\item{accessionDir}{the directory where the accession list can be found}

\item{outDir}{The output directory}
}
\value{

}
\description{
Title
}
