% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-annotation.R
\name{downloadannotation}
\alias{downloadannotation}
\title{Title}
\usage{
downloadannotation(
  maxOrganism = 20,
  accession.list,
  repertoire,
  indextorestart
)
}
\arguments{
\item{indextorestart}{: the index to restart the analysis. for example, if you interupted the analysis after 20 species, you can specify indextorestart=21}
}
\value{

}
\description{
Title
}
