% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.1.gene.annotationbased.R
\name{extract.1.gene.annotationbased}
\alias{extract.1.gene.annotationbased}
\title{Title}
\usage{
extract.1.gene.annotationbased(
  selectedspecies = "ecoli",
  selectedgene = "tolB",
  annotationDir,
  genomeDir,
  outDir
)
}
\arguments{
\item{outDir}{}
}
\value{

}
\description{
Title
}
