% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.sequences.annotation.R
\name{extract.sequences.annotation}
\alias{extract.sequences.annotation}
\title{extrait les sequences comportants colicin dans le produit}
\usage{
extract.sequences.annotation(species, annotationDir, outDir, genomeDir)
}
\arguments{
\item{genomeDir}{}
}
\value{

}
\description{
extrait les sequences comportants colicin dans le produit
}
